Moved user preferences to a separate prefs.py file
